refactor: 更新Spring国际化设计文档，将javax.validation迁移为jakarta.validation包名，调整管理接口路由前缀，优化BeanPostProcessor中的验证器工厂类引用

+++ iot/iot-device-service/src/main/java/io/github/rose/device/vo/ProductCategoryVO.java
@@ -132,5 +132,5 @@ public class ProductCategoryVO {
      * 产品数量
      */
     @Schema(description = "产品数量", example = "10")
-    private Integer productCount;
+    private Long productCount;
 }
